Fix : allow Graphviz to display html label
Graphviz interpret and display a label value as HTML if it is HTML.
See here
https://graphviz.org/docs/attrs/label/

Quoting :
If the value of a label attribute (label for nodes, edges, clusters,
and graphs, and the headlabel and taillabel attributes of an edge) is
given as an HTML string, that is, delimited by <...> rather than
"...", the label is interpreted as an HTML description

-> perform 'a quick guess' (first and last character label values are
'<' and '>' respectively.
If true, do not surround the label value with double quote.

Note : we could use a proper HTML parser but IMHO intentionality is
enough, so I prefer not add dependencies.